Update index.html
updated the game so that it has letters! When the note gets into the zone, you have to type the correct letter to get points. Use space to go up and enter to go down (can't think of a better way...) 
Still a few bugs to work out though... (like for the long notes, if you keep pressing down you get more points)